+++ package.json
@@ -4,16 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "server": "node src/index.js",
-    "test": "jest"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.2",
     "express": "^4.18.2"
-  },
-  "devDependencies": {
-    "jest": "^29.7.0"
   }
 }
