+++ src/index.js
@@ -1,8 +1,6 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const tasksRepository = require("./tasksRepository");
 
-// otro cambio prron
 const app = express();
 const port = 3000;
 
@@ -16,14 +14,13 @@ let tasks = [
 
 // Get all tasks
 app.get('/tasks', (req, res) => {
-  const tasks = tasksRepository.getAll()
   res.json(tasks);
 });
 
 // Get a specific task
 app.get('/tasks/:id', (req, res) => {
   const taskId = parseInt(req.params.id);
-  const task = tasksRepository.getById(taskId)
+  const task = tasks.find((t) => t.id === taskId);
 
   if (task) {
     res.json(task);
@@ -35,7 +32,8 @@ app.get('/tasks/:id', (req, res) => {
 // Create a new task
 app.post('/tasks', (req, res) => {
   const newTask = req.body;
-  tasksRepository.createTask(newTask)
+  newTask.id = tasks.length + 1;
+  tasks.push(newTask);
   res.status(201).json(newTask);
 });
 
@@ -43,10 +41,11 @@ app.post('/tasks', (req, res) => {
 app.put('/tasks/:id', (req, res) => {
   const taskId = parseInt(req.params.id);
   const updatedTask = req.body;
-  const task = tasksRepository.updateTask(taskId, updatedTask)
+  const index = tasks.findIndex((t) => t.id === taskId);
 
-  if (task != null) {
-    res.json(task);
+  if (index !== -1) {
+    tasks[index] = { ...tasks[index], ...updatedTask };
+    res.json(tasks[index]);
   } else {
     res.status(404).json({ error: 'Task not found' });
   }
@@ -55,7 +54,7 @@ app.put('/tasks/:id', (req, res) => {
 // Delete a task
 app.delete('/tasks/:id', (req, res) => {
   const taskId = parseInt(req.params.id);
-  tasksRepository.deleteTask(taskId)
+  tasks = tasks.filter((t) => t.id !== taskId);
   res.sendStatus(204);
 });
